\name{plot.IJROI}
\alias{plot.IJROI}
\title{Plot IJROI object}
\usage{
  \method{plot}{IJROI} (x, add = FALSE, xlab = "",
    ylab = "", main = "", ...)
}
\arguments{
  \item{x}{The IJROI object}

  \item{add}{Whether to add to an existing plot}

  \item{main}{an overall title for the plot: see
  \code{title}}

  \item{xlab}{a title for the x axis: see \code{title}}

  \item{ylab}{a title for the y axis: see \code{title}}

  \item{...}{Additional parameters}
}
\description{
  Plot IJROI object
}
\author{
  David Sterratt
}

