\name{print.ijroi}
\alias{print.ijroi}
\title{Print \code{ijroi} objects}
\usage{
\method{print}{ijroi}(x, all = FALSE, ...)
}
\arguments{
  \item{all}{logical indicating whether to print all
  information from ijroi object as opposed to a subset of
  relevant information. Defaults to \code{FALSE}.}

  \item{x}{\code{ijroi} object to be printed.}

  \item{...}{further arguments passed to
  \code{\link{print}}.}
}
\description{
Print \code{ijroi} objects
}
\author{
Mikko Vihtakari, David Sterratt
}
\seealso{
\code{\link{read.ijroi}}
}

