% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.ijroi.R
\name{write.ijroi}
\alias{write.ijroi}
\title{Write an ImageJ ROI file.}
\usage{
write.ijroi(file, roi, verbose = TRUE)
}
\arguments{
\item{file}{Name of ImageJ ROI file to write}

\item{roi}{A structure of class \code{ijroi} containing the ROI information}

\item{verbose}{Whether to report information}
}
\description{
Write an ImageJ ROI file.
}
\seealso{
\code{\link{read.ijroi}} for reading an ROI file
}
