% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.ijzip.R
\name{write.ijzip}
\alias{write.ijzip}
\title{Write ImageJ zip file containing several ROI files}
\usage{
write.ijzip(file, roi, verbose = TRUE)
}
\arguments{
\item{file}{zip archive to write that will contain a collection of
ImageJ ROI files}

\item{roi}{A list of ROIs}

\item{verbose}{Whether to report information}
}
\description{
Write or add to a zip archive containing ImageJ ROI
  files using the \code{\link{write.ijroi}} function.
}
\seealso{
\code{\link{write.ijroi}}
}
