\name{plot.IJROI}
\alias{plot.IJROI}
\title{Plot IJROI object}
\usage{
  \method{plot}{IJROI} (x, add = FALSE, ...)
}
\arguments{
  \item{x}{The IJROI object}

  \item{add}{Whether to add to an existing plot}

  \item{...}{Additional parameters}
}
\description{
  Plot IJROI object
}
\author{
  David Sterratt
}

