% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.ijroi.R
\name{print.ijroi}
\alias{print.ijroi}
\title{Print \code{ijroi} objects}
\usage{
\method{print}{ijroi}(x, all = FALSE, ...)
}
\arguments{
\item{x}{\code{ijroi} object to be printed.}

\item{all}{logical indicating whether to print all information
from ijroi object as opposed to a subset of relevant information.
Defaults to \code{FALSE}.}

\item{...}{further arguments passed to \code{\link{print}}.}
}
\description{
Print \code{ijroi} objects
}
\seealso{
\code{\link{read.ijroi}}
}
\author{
Mikko Vihtakari, David Sterratt
}
