% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RImageJROI-package.r
\docType{package}
\name{RImageJROI}
\alias{RImageJROI}
\alias{RImageJROI-package}
\title{Read ImageJ Region of Interest (ROI) files}
\description{
Provides functions to read ImageJ (http://imagej.nih.gov/ij/) Region of Interest (ROI) files, to plot the ROIs and to convert them as spatstat (http://spatstat.org/) spatial patterns.
}
\details{
ImageJ ROI objects can be read to R using \code{\link{read.ijroi}} and \code{\link{read.ijzip}} functions. The objects can be plotted using generic \code{\link[=plot.ijroi]{plot}} command and converted to \link{spatstat} spatial patterns by using \code{\link{ij2spatstat}} function.
}

